+++ lib/runners/trace_writer.rb
@@ -7,86 +7,94 @@ module Runners
       @sensitive_strings = sensitive_strings
     end
 
-    def command_line(command_line, recorded_at: Time.now)
-      self << { trace: 'command_line', command_line: command_line.map { |v| masked_string(v) }, recorded_at: recorded_at.utc.iso8601 }
+    def command_line(command_line, recorded_at: now)
+      self << { trace: :command_line, command_line: command_line.map { |v| masked_string(v) }, recorded_at: recorded_at }
     end
 
-    def status(status, recorded_at: Time.now)
-      self << { trace: 'status', status: status.exitstatus, recorded_at: recorded_at.utc.iso8601 }
+    def status(status, recorded_at: now)
+      self << { trace: :status, status: status.exitstatus, recorded_at: recorded_at }
     end
 
-    def stdout(string, recorded_at: Time.now, max_length: 4_000)
+    def stdout(string, recorded_at: now, max_length: 4_000)
       unless string.empty?
-        each_slice(masked_string(string), size: max_length) do |text|
-          self << { trace: 'stdout', string: text, recorded_at: recorded_at.utc.iso8601 }
+        each_slice(masked_string(string), size: max_length) do |text, truncated|
+          self << { trace: :stdout, string: text, recorded_at: recorded_at, truncated: truncated }
         end
       end
     end
 
-    def stderr(string, recorded_at: Time.now, max_length: 4_000)
+    def stderr(string, recorded_at: now, max_length: 4_000)
       unless string.empty?
-        each_slice(masked_string(string), size: max_length) do |text|
-          self << { trace: 'stderr', string: text, recorded_at: recorded_at.utc.iso8601 }
+        each_slice(masked_string(string), size: max_length) do |text, truncated|
+          self << { trace: :stderr, string: text, recorded_at: recorded_at, truncated: truncated }
         end
       end
     end
 
-    def message(message, recorded_at: Time.now, max_length: 4_000, limit: 40_000, omission: "...(truncated)")
+    def message(message, recorded_at: now, max_length: 4_000, limit: 40_000, omission: "...(truncated)")
       string = masked_string(message)
       if string.size > limit
         string = string[0, limit] + omission
+        all_truncated = true
       end
-      each_slice(string, size: max_length) do |text|
-        self << { trace: 'message', message: text, recorded_at: recorded_at.utc.iso8601 }
+
+      block_given = block_given?
+      each_slice(string, size: max_length) do |text, truncated|
+        self << { trace: :message, message: text, recorded_at: recorded_at, truncated: all_truncated || truncated }.tap do |entry|
+          entry[:duration_in_ms] = 0 if block_given
+        end
       end
-      if block_given?
-        start = Time.now
+
+      if block_given
+        start = now
         yield.tap do
-          duration = Time.now - start
-          self << { trace: 'message', message: "    -> #{format_duration_in_secs(duration)}s", recorded_at: (recorded_at + duration).utc.iso8601 }
+          duration = now - start
+          duration_in_ms = (duration * 1000).truncate
+          duration_in_ms = 1 if duration_in_ms == 0 # NOTE: Prevent zero (no time)
+          self << { trace: :message, message: "-> #{duration_in_ms / 1000.0}s", recorded_at: recorded_at + duration,
+                    truncated: false, duration_in_ms: duration_in_ms }
         end
       end
     end
 
-    def header(message, recorded_at: Time.now)
-      self << { trace: 'header', message: masked_string(message), recorded_at: recorded_at.utc.iso8601 }
+    def header(message, recorded_at: now)
+      self << { trace: :header, message: masked_string(message), recorded_at: recorded_at }
     end
 
-    def warning(message, file: nil, recorded_at: Time.now)
-      self << { trace: 'warning', file: file, message: masked_string(message), recorded_at: recorded_at.utc.iso8601 }
+    def warning(message, file: nil, recorded_at: now)
+      self << { trace: :warning, file: file, message: masked_string(message), recorded_at: recorded_at }
     end
 
-    def ci_config(content, recorded_at: Time.now)
-      self << { trace: 'ci_config', content: content, recorded_at: recorded_at.utc.iso8601 }
+    def ci_config(content, file:, recorded_at: now)
+      self << { trace: :ci_config, content: content, file: file, recorded_at: recorded_at }
     end
 
-    def error(message, recorded_at: Time.now, max_length: 4_000)
-      each_slice(masked_string(message), size: max_length) do |text|
-        self << { trace: 'error', message: text, recorded_at: recorded_at.utc.iso8601 }
+    def error(message, recorded_at: now, max_length: 4_000)
+      each_slice(masked_string(message), size: max_length) do |text, truncated|
+        self << { trace: :error, message: text, recorded_at: recorded_at, truncated: truncated }
       end
     end
 
     def <<(object)
+      recorded_at = object[:recorded_at]
+      object = object.merge(recorded_at: recorded_at.utc.iso8601(3)) if recorded_at
       writer << Schema::Trace.anything.coerce(object)
     end
 
     private
+
+    def now
+      Time.now
+    end
 
     def each_slice(string, size:)
       string = string.dup
 
       while string.length > 0
         slice = string.slice!(0, size)
-        if slice.length == size
-          yield(slice + "\\")
-        else
-          yield slice
-        end
+        truncated = slice.length == size
+        yield slice, truncated
       end
-    end
-
-    def format_duration_in_secs(duration)
-      duration.ceil(4).to_s
     end
 
     def masked_string(str)